+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **Initial release**
 - Added `TooltipPlugin` plugin
+- Added `TooltipSet` system set
 - Added `PrimaryTooltip` resource
 - Added `Tooltip` component
 - Added `bevy_reflect` feature